refactor(tui): extract vt100 drawing into tui::render::vt
The VtScreen widget, its style and colour translation, the scrollback
view, and the two legacy fixed-grid screens move into the new render
module with their tests.
