+++ src/tui/render/mod.rs
@@ -1,0 +1,3 @@
+//! Everything that draws: chrome, overlays, modals, and pane contents.
+
+pub(in crate::tui) mod vt;
