+++ src/tui/render/vt.rs
@@ -1,0 +1,297 @@
+//! Drawing a `vt100` screen — and the two legacy fixed-grid views — into a
+//! ratatui buffer.
+
+use ratatui::{
+    Frame,
+    buffer::Buffer,
+    layout::Rect,
+    style::{Color, Modifier, Style},
+    widgets::Widget,
+};
+
+use crate::{
+    screen::SCROLLBACK_LINES,
+    tui::{PaneTextSelection, ScreenCell},
+};
+
+pub(in crate::tui) struct VtScreen<'a> {
+    screen: &'a vt100::Screen,
+    selection: Option<PaneTextSelection>,
+}
+impl<'a> VtScreen<'a> {
+    pub(in crate::tui) fn new(screen: &'a vt100::Screen) -> Self {
+        Self {
+            screen,
+            selection: None,
+        }
+    }
+
+    pub(in crate::tui) fn with_selection(mut self, selection: Option<PaneTextSelection>) -> Self {
+        self.selection = selection;
+        self
+    }
+}
+pub(in crate::tui) fn viewed_screen(screen: &vt100::Screen, scrollback: usize) -> vt100::Screen {
+    let mut screen = screen.clone();
+    screen.set_scrollback(scrollback);
+    screen
+}
+pub(in crate::tui) fn available_scrollback(screen: &vt100::Screen) -> usize {
+    let mut screen = screen.clone();
+    screen.set_scrollback(SCROLLBACK_LINES);
+    screen.scrollback()
+}
+impl Widget for VtScreen<'_> {
+    fn render(self, area: Rect, buf: &mut Buffer) {
+        let (rows, cols) = self.screen.size();
+        for row in 0..rows.min(area.height) {
+            for col in 0..cols.min(area.width) {
+                let Some(source) = self.screen.cell(row, col) else {
+                    continue;
+                };
+                if source.is_wide_continuation() {
+                    continue;
+                }
+                let target = &mut buf[(area.x + col, area.y + row)];
+                let contents = source.contents();
+                target.set_symbol(if contents.is_empty() { " " } else { contents });
+                let style = if self
+                    .selection
+                    .is_some_and(|selection| selection.contains(ScreenCell { row, col }))
+                {
+                    vt_style(source)
+                        .bg(Color::DarkGray)
+                        .add_modifier(Modifier::REVERSED)
+                } else {
+                    vt_style(source)
+                };
+                target.set_style(style);
+            }
+        }
+    }
+}
+fn vt_style(cell: &vt100::Cell) -> Style {
+    let mut modifiers = Modifier::empty();
+    if cell.bold() {
+        modifiers.insert(Modifier::BOLD);
+    }
+    if cell.dim() {
+        modifiers.insert(Modifier::DIM);
+    }
+    if cell.italic() {
+        modifiers.insert(Modifier::ITALIC);
+    }
+    if cell.underline() {
+        modifiers.insert(Modifier::UNDERLINED);
+    }
+    if cell.inverse() {
+        modifiers.insert(Modifier::REVERSED);
+    }
+    Style::default()
+        .fg(vt_color(cell.fgcolor()))
+        .bg(vt_color(cell.bgcolor()))
+        .add_modifier(modifiers)
+}
+fn vt_color(color: vt100::Color) -> Color {
+    match color {
+        vt100::Color::Default => Color::Reset,
+        vt100::Color::Idx(index) => Color::Indexed(index),
+        vt100::Color::Rgb(red, green, blue) => Color::Rgb(red, green, blue),
+    }
+}
+pub(in crate::tui) fn render_guest_screen(
+    frame: &mut Frame<'_>,
+    screen: &vt100::Screen,
+    footer: &str,
+) {
+    let area = frame.area();
+    let screen_height = screen.size().0.min(area.height.saturating_sub(1));
+    let screen_area = Rect::new(area.x, area.y, area.width, screen_height);
+    frame.render_widget(VtScreen::new(screen), screen_area);
+    let (row, col) = screen.cursor_position();
+    if !screen.hide_cursor() && row < screen_height && col < area.width {
+        frame.set_cursor_position((area.x + col, area.y + row));
+    }
+    if area.height > 0 {
+        let footer_y = area.y + screen_area.height;
+        frame
+            .buffer_mut()
+            .set_string(area.x, footer_y, footer, Style::default());
+    }
+}
+pub(in crate::tui) fn render_host_screen(
+    frame: &mut Frame<'_>,
+    screen: &vt100::Screen,
+    footer: &str,
+) {
+    render_guest_screen(frame, screen, footer);
+}
+
+#[cfg(test)]
+mod tests {
+    use ratatui::{
+        Terminal,
+        backend::TestBackend,
+        style::{Color, Modifier},
+    };
+
+    use crate::{
+        screen::{GuestScreen, HostScreen},
+        tui::{PaneTextSelection, ScreenCell},
+    };
+
+    use super::{VtScreen, render_guest_screen, viewed_screen};
+
+    #[test]
+    fn remote_renderer_keeps_host_grid_fixed_and_draws_a_footer() {
+        let mut host = HostScreen::new(1, 3).expect("host screen");
+        let frame = host.process_pty(b"abc").expect("frame");
+        let mut guest = GuestScreen::new();
+        guest
+            .apply_snapshot(frame.sequence, &frame.snapshot)
+            .expect("snapshot");
+        let mut terminal = Terminal::new(TestBackend::new(5, 3)).expect("test terminal");
+        terminal
+            .draw(|frame| {
+                render_guest_screen(
+                    frame,
+                    guest.screen().expect("guest screen"),
+                    "controller: abcdef idle",
+                )
+            })
+            .expect("render");
+        let buffer = terminal.backend().buffer();
+        assert_eq!(buffer[(0, 0)].symbol(), "a");
+        assert_eq!(buffer[(2, 0)].symbol(), "c");
+        assert_eq!(buffer[(3, 0)].symbol(), " ");
+        assert_eq!(buffer[(0, 1)].symbol(), "c");
+        assert_eq!(buffer[(0, 2)].symbol(), " ");
+    }
+
+    #[test]
+    fn remote_renderer_places_the_host_cursor() {
+        let mut host = HostScreen::new(1, 3).expect("host screen");
+        let frame = host.process_pty(b"ab").expect("frame");
+        let mut guest = GuestScreen::new();
+        guest
+            .apply_snapshot(frame.sequence, &frame.snapshot)
+            .expect("snapshot");
+        let mut terminal = Terminal::new(TestBackend::new(5, 3)).expect("test terminal");
+
+        terminal
+            .draw(|frame| {
+                render_guest_screen(
+                    frame,
+                    guest.screen().expect("guest screen"),
+                    "controller: peer typing",
+                );
+            })
+            .expect("render");
+
+        terminal.backend_mut().assert_cursor_position((2, 0));
+    }
+
+    #[test]
+    fn renders_vt100_cell_styles() {
+        let mut parser = vt100::Parser::new(1, 3, 0);
+        parser.process(b"\x1b[31;44;1mX");
+        let mut terminal = Terminal::new(TestBackend::new(3, 1)).expect("test terminal");
+        terminal
+            .draw(|frame| frame.render_widget(VtScreen::new(parser.screen()), frame.area()))
+            .expect("render should work");
+        let buffer = terminal.backend().buffer();
+
+        assert_eq!(buffer[(0, 0)].symbol(), "X");
+        assert_eq!(buffer[(0, 0)].fg, Color::Indexed(1));
+        assert_eq!(buffer[(0, 0)].bg, Color::Indexed(4));
+        assert!(buffer[(0, 0)].modifier.contains(Modifier::BOLD));
+    }
+
+    #[test]
+    fn renderer_highlights_stream_selected_cells() {
+        let mut parser = vt100::Parser::new(3, 4, 0);
+        parser.process(b"abcd\r\nefgh\r\nijkl");
+        let selection = PaneTextSelection {
+            pane_id: 1,
+            anchor: ScreenCell { row: 2, col: 1 },
+            cursor: ScreenCell { row: 0, col: 2 },
+        };
+        let mut terminal = Terminal::new(TestBackend::new(4, 3)).expect("test terminal");
+
+        terminal
+            .draw(|frame| {
+                frame.render_widget(
+                    VtScreen::new(parser.screen()).with_selection(Some(selection)),
+                    frame.area(),
+                )
+            })
+            .expect("render should work");
+
+        let buffer = terminal.backend().buffer();
+        assert_ne!(buffer[(1, 0)].bg, Color::DarkGray);
+        assert_eq!(buffer[(2, 0)].bg, Color::DarkGray);
+        assert_eq!(buffer[(0, 1)].bg, Color::DarkGray);
+        assert_eq!(buffer[(1, 2)].bg, Color::DarkGray);
+        assert_ne!(buffer[(2, 2)].bg, Color::DarkGray);
+        assert!(buffer[(2, 0)].modifier.contains(Modifier::REVERSED));
+    }
+
+    #[test]
+    fn renderer_keeps_the_parser_grid_fixed() {
+        let mut parser = vt100::Parser::new(2, 3, 0);
+        parser.process(b"abc\r\ndef");
+        let mut terminal = Terminal::new(TestBackend::new(5, 4)).expect("test terminal");
+        terminal
+            .draw(|frame| frame.render_widget(VtScreen::new(parser.screen()), frame.area()))
+            .expect("render should work");
+        let buffer = terminal.backend().buffer();
+
+        assert_eq!(buffer[(0, 0)].symbol(), "a");
+        assert_eq!(buffer[(2, 1)].symbol(), "f");
+        assert_eq!(buffer[(3, 0)].symbol(), " ");
+        assert_eq!(buffer[(0, 2)].symbol(), " ");
+    }
+
+    #[test]
+    fn renderer_uses_the_requested_scrollback_view() {
+        let mut parser = vt100::Parser::new(1, 3, 10);
+        parser.process(b"one\r\ntwo");
+        let live = parser.screen().clone();
+        let history = viewed_screen(parser.screen(), 1);
+        assert_eq!(history.scrollback(), 1);
+        assert_ne!(
+            history.cell(0, 0).expect("history cell").contents(),
+            live.cell(0, 0).expect("live cell").contents()
+        );
+        let mut terminal = Terminal::new(TestBackend::new(3, 1)).expect("test terminal");
+
+        terminal
+            .draw(|frame| frame.render_widget(VtScreen::new(&history), frame.area()))
+            .expect("render should work");
+
+        assert_eq!(
+            terminal.backend().buffer()[(0, 0)].symbol(),
+            history.cell(0, 0).expect("history cell").contents()
+        );
+    }
+
+    #[test]
+    fn renderer_erases_cells_cleared_by_the_pty() {
+        let mut parser = vt100::Parser::new(1, 3, 0);
+        let mut terminal = Terminal::new(TestBackend::new(3, 1)).expect("test terminal");
+        parser.process(b"abc");
+        terminal
+            .draw(|frame| frame.render_widget(VtScreen::new(parser.screen()), frame.area()))
+            .expect("initial render");
+
+        parser.process(b"\x1b[2J\x1b[H");
+        terminal
+            .draw(|frame| frame.render_widget(VtScreen::new(parser.screen()), frame.area()))
+            .expect("clear render");
+
+        let buffer = terminal.backend().buffer();
+        assert_eq!(buffer[(0, 0)].symbol(), " ");
+        assert_eq!(buffer[(1, 0)].symbol(), " ");
+        assert_eq!(buffer[(2, 0)].symbol(), " ");
+    }
+}
